Authentication and Authorization into feast auth module.
