+++ auth/src/main/java/feast/auth/authorization/AuthorizationProvider.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package feast.core.auth.authorization;
+package feast.auth.authorization;
 
 import org.springframework.security.core.Authentication;
