+++ common/pom.xml
@@ -56,6 +56,11 @@
             <artifactId>lombok</artifactId>
         </dependency>
         <dependency>
+			<groupId>javax.validation</groupId>
+			<artifactId>validation-api</artifactId>
+			<version>2.0.0.Final</version>
+		</dependency>
+        <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
             <version>4.12</version>
